+++ include/tropicssl/base64.h
@@ -35,6 +35,8 @@
 #ifndef TROPICSSL_BASE64_H
 #define TROPICSSL_BASE64_H
 
+#include <string.h>
+
 #define TROPICSSL_ERR_BASE64_BUFFER_TOO_SMALL               -0x0010
 #define TROPICSSL_ERR_BASE64_INVALID_CHARACTER              -0x0012
 
@@ -57,8 +59,8 @@ extern "C" {
 	 * \note           Call this function with *dlen = 0 to obtain the
 	 *                 required buffer size in *dlen
 	 */
-	int base64_encode(unsigned char *dst, int *dlen,
-			  const unsigned char *src, int slen);
+	int base64_encode(unsigned char *dst, size_t *dlen,
+			  const unsigned char *src, size_t slen);
 
 	/**
 	 * \brief          Decode a base64-formatted buffer
@@ -76,8 +78,8 @@ extern "C" {
 	 * \note           Call this function with *dlen = 0 to obtain the
 	 *                 required buffer size in *dlen
 	 */
-	int base64_decode(unsigned char *dst, int *dlen,
-			  const unsigned char *src, int slen);
+	int base64_decode(unsigned char *dst, size_t *dlen,
+			  const unsigned char *src, size_t slen);
 
 	/**
 	 * \brief          Checkup routine
