+++ include/tropicssl/md2.h
@@ -35,6 +35,8 @@
 #ifndef TROPICSSL_MD2_H
 #define TROPICSSL_MD2_H
 
+#include <string.h>
+
 /**
  * \brief          MD2 context structure
  */
@@ -45,7 +47,7 @@ typedef struct {
 
 	unsigned char ipad[64];	/*!< HMAC: inner padding        */
 	unsigned char opad[64];	/*!< HMAC: outer padding        */
-	int left;		/*!< amount of data in buffer   */
+	size_t left;			/*!< amount of data in buffer   */
 } md2_context;
 
 #ifdef __cplusplus
@@ -66,7 +68,7 @@ extern "C" {
 	 * \param input    buffer holding the  data
 	 * \param ilen     length of the input data
 	 */
-	void md2_update(md2_context * ctx, const unsigned char *input, int ilen);
+	void md2_update(md2_context * ctx, const unsigned char *input, size_t ilen);
 
 	/**
 	 * \brief          MD2 final digest
@@ -83,7 +85,7 @@ extern "C" {
 	 * \param ilen     length of the input data
 	 * \param output   MD2 checksum result
 	 */
-	void md2(const unsigned char *input, int ilen, unsigned char output[16]);
+	void md2(const unsigned char *input, size_t ilen, unsigned char output[16]);
 
 	/**
 	 * \brief          Output = MD2( file contents )
@@ -103,7 +105,7 @@ extern "C" {
 	 * \param key      HMAC secret key
 	 * \param keylen   length of the HMAC key
 	 */
-	void md2_hmac_starts(md2_context * ctx, const unsigned char *key, int keylen);
+	void md2_hmac_starts(md2_context * ctx, const unsigned char *key, size_t keylen);
 
 	/**
 	 * \brief          MD2 HMAC process buffer
@@ -112,7 +114,7 @@ extern "C" {
 	 * \param input    buffer holding the  data
 	 * \param ilen     length of the input data
 	 */
-	void md2_hmac_update(md2_context * ctx, const unsigned char *input, int ilen);
+	void md2_hmac_update(md2_context * ctx, const unsigned char *input, size_t ilen);
 
 	/**
 	 * \brief          MD2 HMAC final digest
@@ -131,8 +133,8 @@ extern "C" {
 	 * \param ilen     length of the input data
 	 * \param output   HMAC-MD2 result
 	 */
-	void md2_hmac(const unsigned char *key, int keylen,
-		      const unsigned char *input, int ilen, unsigned char output[16]);
+	void md2_hmac(const unsigned char *key, size_t keylen,
+		      const unsigned char *input, size_t ilen, unsigned char output[16]);
 
 	/**
 	 * \brief          Checkup routine
