+++ include/tropicssl/md5.h
@@ -35,12 +35,14 @@
 #ifndef TROPICSSL_MD5_H
 #define TROPICSSL_MD5_H
 
+#include <string.h>
+#include <inttypes.h>
 /**
  * \brief          MD5 context structure
  */
 typedef struct {
-	unsigned long total[2];	/*!< number of bytes processed  */
-	unsigned long state[4];	/*!< intermediate digest state  */
+	uint32_t total[2];	/*!< number of bytes processed  */
+	uint32_t state[4];	/*!< intermediate digest state  */
 	unsigned char buffer[64];	/*!< data block being processed */
 
 	unsigned char ipad[64];	/*!< HMAC: inner padding        */
@@ -65,7 +67,7 @@ extern "C" {
 	 * \param input    buffer holding the  data
 	 * \param ilen     length of the input data
 	 */
-	void md5_update(md5_context * ctx, const unsigned char *input, int ilen);
+	void md5_update(md5_context * ctx, const unsigned char *input, size_t ilen);
 
 	/**
 	 * \brief          MD5 final digest
@@ -82,7 +84,7 @@ extern "C" {
 	 * \param ilen     length of the input data
 	 * \param output   MD5 checksum result
 	 */
-	void md5(const unsigned char *input, int ilen, unsigned char output[16]);
+	void md5(const unsigned char *input, size_t ilen, unsigned char output[16]);
 
 	/**
 	 * \brief          Output = MD5( file contents )
@@ -102,7 +104,7 @@ extern "C" {
 	 * \param key      HMAC secret key
 	 * \param keylen   length of the HMAC key
 	 */
-	void md5_hmac_starts(md5_context * ctx, const unsigned char *key, int keylen);
+	void md5_hmac_starts(md5_context * ctx, const unsigned char *key, size_t keylen);
 
 	/**
 	 * \brief          MD5 HMAC process buffer
@@ -111,7 +113,7 @@ extern "C" {
 	 * \param input    buffer holding the  data
 	 * \param ilen     length of the input data
 	 */
-	void md5_hmac_update(md5_context * ctx, const unsigned char *input, int ilen);
+	void md5_hmac_update(md5_context * ctx, const unsigned char *input, size_t ilen);
 
 	/**
 	 * \brief          MD5 HMAC final digest
@@ -130,8 +132,8 @@ extern "C" {
 	 * \param ilen     length of the input data
 	 * \param output   HMAC-MD5 result
 	 */
-	void md5_hmac(const unsigned char *key, int keylen,
-		      const unsigned char *input, int ilen, unsigned char output[16]);
+	void md5_hmac(const unsigned char *key, size_t keylen,
+		      const unsigned char *input, size_t ilen, unsigned char output[16]);
 
 	/**
 	 * \brief          Checkup routine
