+++ include/tropicssl/net.h
@@ -35,6 +35,8 @@
 #ifndef TROPICSSL_NET_H
 #define TROPICSSL_NET_H
 
+#include <string.h>
+
 #define TROPICSSL_ERR_NET_UNKNOWN_HOST                      -0x0F00
 #define TROPICSSL_ERR_NET_SOCKET_FAILED                     -0x0F10
 #define TROPICSSL_ERR_NET_CONNECT_FAILED                    -0x0F20
@@ -110,7 +112,7 @@ extern "C" {
 	 *                 or a negative error code; TROPICSSL_ERR_NET_TRY_AGAIN
 	 *                 indicates read() is blocking.
 	 */
-	int net_recv(void *ctx, unsigned char *buf, int len);
+	int net_recv(void *ctx, unsigned char *buf, size_t len);
 
 	/**
 	 * \brief          Write at most 'len' characters. len is updated to
@@ -120,7 +122,7 @@ extern "C" {
 	 *                 or a negative error code; TROPICSSL_ERR_NET_TRY_AGAIN
 	 *                 indicates write() is blocking.
 	 */
-	int net_send(void *ctx, unsigned char *buf, int len);
+	int net_send(void *ctx, unsigned char *buf, size_t len);
 
 	/**
 	 * \brief          Gracefully shutdown the connection
