+++ include/tropicssl/ssl.h
@@ -172,7 +172,7 @@ typedef struct _ssl_context ssl_context;
 struct _ssl_session {
 	time_t start;		/*!< starting time      */
 	int cipher;		/*!< chosen cipher      */
-	int length;		/*!< session id length  */
+	size_t length;		/*!< session id length  */
 	unsigned char id[32];	/*!< session identifier */
 	unsigned char master[48];	/*!< the master secret  */
 	ssl_session *next;	/*!< next session entry */
@@ -195,8 +195,8 @@ struct _ssl_context {
 	 */
 	int (*f_rng) (void *);
 	void (*f_dbg) (void *, int, const char *);
-	int (*f_recv) (void *, unsigned char *, int);
-	int (*f_send) (void *, unsigned char *, int);
+	int (*f_recv) (void *, unsigned char *, size_t);
+	int (*f_send) (void *, unsigned char *, size_t);
 
 	void *p_rng;		/*!< context for the RNG function     */
 	void *p_dbg;		/*!< context for the debug function   */
@@ -220,12 +220,12 @@ struct _ssl_context {
 	unsigned char *in_msg;	/*!< the message contents (in_hdr+5)  */
 	unsigned char *in_offt;	/*!< read offset in application data  */
 
-	int in_msgtype;		/*!< record header: message type      */
-	int in_msglen;		/*!< record header: message length    */
-	int in_left;		/*!< amount of data read so far       */
+	int in_msgtype;			/*!< record header: message type      */
+	size_t in_msglen;		/*!< record header: message length    */
+	size_t in_left;			/*!< amount of data read so far       */
 
-	int in_hslen;		/*!< current handshake message length */
-	int nb_zero;		/*!< # of 0-length encrypted messages */
+	size_t in_hslen;		/*!< current handshake message length */
+	int nb_zero;			/*!< # of 0-length encrypted messages */
 
 	/*
 	 * Record layer (outgoing data)
@@ -234,9 +234,9 @@ struct _ssl_context {
 	unsigned char *out_hdr;	/*!< 5-byte record header (out_ctr+8) */
 	unsigned char *out_msg;	/*!< the message contents (out_hdr+5) */
 
-	int out_msgtype;	/*!< record header: message type      */
-	int out_msglen;		/*!< record header: message length    */
-	int out_left;		/*!< amount of data not yet written   */
+	int out_msgtype;		/*!< record header: message type      */
+	size_t out_msglen;		/*!< record header: message length    */
+	size_t out_left;		/*!< amount of data not yet written   */
 
 	/*
 	 * PKI layer
@@ -261,11 +261,11 @@ struct _ssl_context {
 
 	int do_crypt;		/*!<  en(de)cryption flag     */
 	int *ciphers;		/*!<  allowed ciphersuites    */
-	int pmslen;		/*!<  premaster length        */
-	int keylen;		/*!<  symmetric key length    */
-	int minlen;		/*!<  min. ciphertext length  */
-	int ivlen;		/*!<  IV length               */
-	int maclen;		/*!<  MAC length              */
+	size_t pmslen;		/*!<  premaster length        */
+	unsigned int keylen;		/*!<  symmetric key length    */
+	size_t minlen;		/*!<  min. ciphertext length  */
+	size_t ivlen;		/*!<  IV length               */
+	size_t maclen;		/*!<  MAC length              */
 
 	unsigned char randbytes[64];	/*!<  random bytes            */
 	unsigned char premaster[256];	/*!<  premaster secret        */
@@ -276,14 +276,14 @@ struct _ssl_context {
 	unsigned char mac_enc[32];	/*!<  MAC (encryption)        */
 	unsigned char mac_dec[32];	/*!<  MAC (decryption)        */
 
-	unsigned long ctx_enc[128];	/*!<  encryption context      */
-	unsigned long ctx_dec[128];	/*!<  decryption context      */
+	uint32_t ctx_enc[128];	/*!<  encryption context      */
+	uint32_t ctx_dec[128];	/*!<  decryption context      */
 
 	/*
 	 * TLS extensions
 	 */
 	unsigned char *hostname;
-	unsigned long hostname_len;
+	size_t hostname_len;
 };
 
 #ifdef __cplusplus
@@ -357,9 +357,9 @@ extern "C" {
 	 * \param p_send   write parameter
 	 */
 	void ssl_set_bio(ssl_context * ssl,
-			 int (*f_recv) (void *, unsigned char *, int),
+			 int (*f_recv) (void *, unsigned char *, size_t),
 			 void *p_recv, int (*f_send) (void *, unsigned char *,
-						      int), void *p_send);
+						      size_t), void *p_send);
 
 	/**
 	 * \brief          Set the session callbacks (server-side only)
@@ -443,7 +443,7 @@ extern "C" {
 	 *
 	 * \return         how many bytes are available in the read buffer
 	 */
-	int ssl_get_bytes_avail(const ssl_context * ssl);
+	size_t ssl_get_bytes_avail(const ssl_context * ssl);
 
 	/**
 	 * \brief          Return the result of the certificate verification
@@ -487,7 +487,7 @@ extern "C" {
 	 * \return         This function returns the number of bytes read,
 	 *                 or a negative error code.
 	 */
-	int ssl_read(ssl_context * ssl, unsigned char *buf, int len);
+	int ssl_read(ssl_context * ssl, unsigned char *buf, size_t len);
 
 	/**
 	 * \brief          Write exactly 'len' application data bytes
@@ -503,7 +503,7 @@ extern "C" {
 	 *                 it must be called later with the *same* arguments,
 	 *                 until it returns a positive value.
 	 */
-	int ssl_write(ssl_context * ssl, const unsigned char *buf, int len);
+	int ssl_write(ssl_context * ssl, const unsigned char *buf, size_t len);
 
 	/**
 	 * \brief          Notify the peer that the connection is being closed
@@ -525,7 +525,7 @@ extern "C" {
 	void ssl_calc_verify(ssl_context * ssl, unsigned char hash[36]);
 
 	int ssl_read_record(ssl_context * ssl);
-	int ssl_fetch_input(ssl_context * ssl, int nb_want);
+	int ssl_fetch_input(ssl_context * ssl, size_t nb_want);
 
 	int ssl_write_record(ssl_context * ssl);
 	int ssl_flush_output(ssl_context * ssl);
