+++ library/padlock.c
@@ -79,9 +79,9 @@ int padlock_xcryptecb(aes_context * ctx,
 		      const unsigned char input[16], unsigned char output[16])
 {
 	int ebx;
-	unsigned long *rk;
-	unsigned long *blk;
-	unsigned long *ctrl;
+	uint32_t *rk;
+	uint32_t *blk;
+	uint32_t *ctrl;
 	unsigned char buf[256];
 
 	rk = ctx->rk;
@@ -110,9 +110,9 @@ int padlock_xcryptcbc(aes_context * ctx,
 		      const unsigned char *input, unsigned char *output)
 {
 	int ebx, count;
-	unsigned long *rk;
-	unsigned long *iw;
-	unsigned long *ctrl;
+	uint32_t *rk;
+	uint32_t *iw;
+	uint32_t *ctrl;
 	unsigned char buf[256];
 
 	if (((long)input & 15) != 0 || ((long)output & 15) != 0)
